Assorted styling improvements
Change symbolizer ordering to be based of style order in layer, not
order in file
Properly apply opacity rules
Add text-transform

+++ WhirlyGlobeSrc/WhirlyGlobe-MaplyComponent/src/MapnikStyleRule.m
@@ -88,8 +88,10 @@
 
 
 - (NSString*)description {
-  return [NSString stringWithFormat:@"%lu - %lu(%lu - %lu): %@, %d symbolizers", (unsigned long)self.minScaleDenominator,
-          (unsigned long)self.maxScaleDenomitator, (unsigned long)self.minZoom, (unsigned long)self.maxZoom, self.filterPredicate, self.symbolizers.count];
+  return [NSString stringWithFormat:@"%lu - %lu(%lu - %lu): %@, %lu symbolizers",
+          (unsigned long)self.minScaleDenominator, (unsigned long)self.maxScaleDenomitator,
+          (unsigned long)self.minZoom, (unsigned long)self.maxZoom, self.filterPredicate,
+          (unsigned long)self.symbolizers.count];
 }
